requirements.txt
flask
flask-session
mysql-connector-python
werkzeug
openai
